refactor(tests): remove design-fix-test-plan skill and update related tests
### Summary

- Deleted the `design-fix-test-plan` skill and its associated documentation files (`README.md` and `SKILL.md`).
- Refactored E2E tests for binary and file fields to improve structure and maintainability by consolidating import statements and enhancing URL matching logic.
- Introduced a new helper file for test data to streamline test interactions across binary and file fields.

### Checklist
- [x] Removed obsolete skill files
- [x] Improved test readability and maintainability

This commit cleans up the codebase by removing unused components and enhancing the clarity of existing tests.
