+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-file-field/components/dot-file-field/dot-file-field.component.ts
@@ -279,6 +279,7 @@ export class DotFileFieldComponent
             header,
             appendTo: 'body',
             closable: true,
+            closeAriaLabel: 'Close',
             closeOnEscape: false,
             draggable: false,
             keepInViewport: false,
